test: 엔티티 클래스, 리포지토리  테스트 분리
