+++ src/main/java/qna/domain/DeleteHistory.java
@@ -38,7 +38,7 @@ public class DeleteHistory {
         return Objects.equals(id, that.id) &&
                 contentType == that.contentType &&
                 Objects.equals(contentId, that.contentId) &&
-                Objects.equals(deletedBy, that.deletedBy);
+                Objects.equals(deletedBy.getId(), that.deletedBy.getId());
     }
 
     @Override
