+++ src/main/java/qna/domain/Question.java
@@ -46,7 +46,7 @@ public class Question extends BaseEntity {
     }
 
     public boolean isOwner(User writer) {
-        return this.writer.equals(writer);
+        return this.writer.getId().equals(writer.getId());
     }
 
     public void addAnswer(Answer answer) {
